Big brain implemented!
