+++ src/ai/components/approach_player_components.rs
@@ -1,0 +1,15 @@
+use bevy::prelude::{Component, Entity};
+use big_brain::prelude::{ActionBuilder, Actor, ScorerBuilder};
+
+#[derive(Clone, Component, Debug, Default)]
+pub struct ApproachPlayerData {
+    pub seen_player: Option<Entity>,
+}
+
+// Scorers are the same as in the thirst example.
+#[derive(Clone, Component, Debug, ScorerBuilder)]
+pub struct ApproachPlayerScore;
+
+/// An action where the actor moves to the closest water source
+#[derive(Clone, Component, Debug, ActionBuilder)]
+pub struct ApproachPlayerAction {}
