+++ src/ai/components/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod avoid_wall_components;
+pub(crate) mod move_forward_components;
+pub(crate) mod approach_player_components;
