+++ src/ai/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod components;
+pub(crate) mod systems;
