+++ src/ai/systems/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod approach_player_systems;
+pub(crate) mod avoid_walls_systems;
+pub(crate) mod move_forward_systems;
