+++ src/enemy/components/mod.rs
@@ -1,2 +1,1 @@
 pub(crate) mod general;
-pub(crate) mod bonsai_ai_components;
