+++ src/enemy/systems/mod.rs
@@ -1,4 +1,1 @@
 pub(crate) mod spawn_aliens;
-pub(crate) mod bonsai_ai_systems;
-pub(crate) mod alien_sight;
-pub(crate) mod big_brain_ai_systems;
